Move farmer to module, missing icon
