+++ src/main/resources/assets/minecolonies/gui/layouthuts/layoutfarmerfields.xml
@@ -1,0 +1,19 @@
+<window size="190 244" pause="false" lightbox="false">
+    <image source="minecolonies:textures/gui/builderhut/builder_paper.png" size="190 244"/>
+    <label size="130 11" pos="30 14" color="black" textalign="MIDDLE"
+           label="$(com.minecolonies.coremod.gui.workerhuts.farmerhut.fields):"/>
+    <label id="AssignmentModeLabel" size="164 11" pos="13 29" color="black" textalign="MIDDLE"
+           label="$(com.minecolonies.coremod.gui.workerhuts.farmerhut.assignmentmode)"/>
+    <button id="assignmentMode" size="60 15" pos="65 42"
+            label="$(com.minecolonies.coremod.gui.hiring.off)"/>
+
+    <list id="fields" size="164 154" pos="13 60">
+        <box size="100% 34" linewidth="2">
+            <itemicon id="icon" size="16 16" pos="1 1"/>
+            <label id="worker" color="black" pos="20 2" size="130 9"/>
+            <label id="dist" color="black" pos="20 12" size="65 9"/>
+            <label id="dir" color="black" pos="20 22" size="65 9"/>
+            <button id="assignFarm" pos="100 16" size="50 15"/>
+        </box>
+    </list>
+</window>
