--- src/main/resources/assets/minecolonies/gui/windowhutfarmer.xml
@@ -1,31 +1,0 @@
-<window xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" size="190 244" pause="false"
-        lightbox="false"
-        xsi:noNamespaceSchemaLocation="file:../../../../../api/java/com/minecolonies/blockout/blockOut.xsd">
-    <switch id="pages" size="100% 100%">
-        <view id="pageActions" size="100% 100%">
-            <layout source="minecolonies:gui/layouthuts/layouthutpageactions.xml"/>
-        </view>
-        <view id="pageFields" size="100% 100%">
-            <image source="minecolonies:textures/gui/builderhut/builder_paper.png" size="190 244"/>
-            <label size="130 11" pos="30 14" color="black" textalign="MIDDLE"
-                   label="$(com.minecolonies.coremod.gui.workerhuts.farmerhut.fields):"/>
-            <label id="AssignmentModeLabel" size="164 11" pos="13 29" color="black" textalign="MIDDLE"
-                   label="$(com.minecolonies.coremod.gui.workerhuts.farmerhut.assignmentmode)"/>
-            <button id="assignmentMode" size="60 15" pos="65 42"
-                    label="$(com.minecolonies.coremod.gui.hiring.off)"/>
-
-            <list id="fields" size="164 154" pos="13 60">
-                <box size="100% 34" linewidth="2">
-                    <itemicon id="icon" size="16 16" pos="1 1"/>
-                    <label id="worker" color="black" pos="20 2" size="130 9"/>
-                    <label id="dist" color="black" pos="20 12" size="65 9"/>
-                    <label id="dir" color="black" pos="20 22" size="65 9"/>
-                    <button id="assignFarm" pos="100 16" size="50 15"/>
-                </box>
-            </list>
-        </view>
-    </switch>
-    <buttonimage id="prevPage" size="18 10" pos="13 13" source="minecolonies:textures/gui/builderhut/turn_page_left.png"/>
-    <buttonimage id="nextPage" size="18 10" pos="159 13" source="minecolonies:textures/gui/builderhut/turn_page_right.png"/>
-    <label id="pageNum" size="20 11" pos="158 222" textalign="MIDDLE" label="" color="black"/>
-</window>
